Move the GDN kernels under kernels/attention
GDN is the linear-attention core of qwen3-next; house it with the other
attention kernels and mirror the upstream maxtext layout
(kernels/attention/gated_delta_network.py) so the two implementations
stay diffable side by side.
