+++ src/megatext/kernels/attention/gated_delta_network.py
